+++ CHANGELOG.md
@@ -5,8 +5,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.5.9-11] - 2020-06-25
 ### Added
 - Counter: total count (lines and papers)
+- Collecter class
+- Compiler: contents title
 ### Changed
 - Compiler: outline comment as a special format
 ### Fixed
@@ -523,7 +527,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Example story as usage.
 - Output story as markdown.
 
-[Unreleased]: https://github.com/nagisc007/storybuilder/compare/v0.5.9-10...HEAD
+[Unreleased]: https://github.com/nagisc007/storybuilder/compare/v0.5.9-11...HEAD
+[0.5.9-11]: https://github.com/nagisc007/storybuilder/releases/v0.5.9-11
 [0.5.9-10]: https://github.com/nagisc007/storybuilder/releases/v0.5.9-10
 [0.5.9-9]: https://github.com/nagisc007/storybuilder/releases/v0.5.9-9
 [0.5.9-8]: https://github.com/nagisc007/storybuilder/releases/v0.5.9-8